cinemaslides: Handle filenames with spaces and rogue chars. Glob dirs.
Basic idea is to replace problematic filenames with temporary links
which will behave nicely throughout the pipeline run (a mix of shell
calls and internal handling).

The mechanism is not fully transparent yet as rogue members of ARGV will
not show up in logger reporting (the temporary links will) which should
be fairly confusing. See the follow-up commit for the changes wrt
@source and @source handling.

Also, if you must, you can now throw directories with subdirs in them at
cinemaslides and it will glob recursively through those.

Thanks, Andrae Steiner, for discussion and feedback.